完善代码增加保存接口

+++ eureka-client/src/main/java/com/suyou/eurekaclient/entity/SysPermissionEntity.java
@@ -23,7 +23,7 @@ public class SysPermissionEntity implements Serializable {
 	 * 
 	 */
 	@TableId
-	private Long oid;
+	private int oid;
 	/**
 	 * 
 	 */
