+++ eureka-client/src/main/java/com/suyou/eurekaclient/entity/SysRoleEntity.java
@@ -23,7 +23,7 @@ public class SysRoleEntity implements Serializable {
 	 * 
 	 */
 	@TableId
-	private Long oid;
+	private int oid;
 	/**
 	 * 
 	 */
