+++ eureka-client/src/main/java/com/suyou/eurekaclient/entity/SysRolePermissionEntity.java
@@ -23,7 +23,7 @@ public class SysRolePermissionEntity implements Serializable {
 	 * 
 	 */
 	@TableId
-	private Long oid;
+	private int oid;
 	/**
 	 * 
 	 */
